+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- next-header -->
 
 ## [Unreleased] - ReleaseDate
+
+## [0.1.2] - 2025-03-18
 - Add some flags to `virtiofsd` call to improve performance and decrease guest memory usage
 - Fix multiple `-v` options not working due to incorrect fstab getting generated
 - Add virtio-pmem device support with `--pmem`
@@ -19,6 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - First somewhat working version
 
 <!-- next-url -->
-[Unreleased]: https://github.com/svenstaro/vmexec/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/svenstaro/vmexec/compare/v0.1.2...HEAD
+[0.1.2]: https://github.com/svenstaro/vmexec/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/svenstaro/vmexec/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/svenstaro/dummyhttp/compare/v0.1.0...v0.1.0